create allowance button component
